NightMode Styling
Anki 2.1.20 recently added a native NightMode setting in the config. It works by automatically inveryting most colors, however it seems to not work for specifically set rgba values, tho it provides a ".nightMode" override for the css to handle things like that. 

This PR contains the styling necessary to have all things be visible in that mode. I tweeked the alpha variables a bit to what looks good to me personally, feel free to change it of course 👍