Redesign and make responsive
Main changes:

- Text and images now scale better to devices with large screens.
    - This is not infinite though: there is a max width to keep the card
      components close to one another.
- The pictures (pegs) are no longer blurry.
- The periodic table is larger and the position of the element under
  consideration easier to see.
- Hopefully, the cards are also a bit more visually pleasing, but that
  wasn't the main goal. Anyway, this is quite subjective, so I've been
  quite conservative.

I considered making the layout of the periodic table change based on
viewport width, to fit the f-block (lanthanides and actinides) alongside
the other blocks, but it didn't seem worth it. While it wouldn't have
been super hard to do, it would make the table really wide and this
would require changes to the rest of the card layout (move the picture
to another row, and perhaps change the text layout for the description).